pyproject.toml
[tool.poetry]
name = "Quality-Assurance-QA"
version = "0.1.0"
description = ""
authors = ["Bruno Gomes QA <bruno.gomes.qa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
flask = "3.0.3"
blue = "^0.9.1"
flask-pydantic-spec = "^0.6.0"
sqlalchemy = "<2.0"
cryptography = "^3.4.7"
taskipy = "^1.12.2"
python-dotenv = "^1.0.1"
pymysql = "1.1.0"
pytest = "^8.2.1"
streamlit = "^1.36.0"
requests = "^2.32.3"
streamlit-pydantic = "^0.6.0"
streamlit-js-eval = "^0.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
